feat(lib): parse tags array from string
Parse tags array from string if not already an array. it is possible
for the tags array to actually be a string. In the case of being passed
from the command line tool `--dockerTags=x` yargs will only coerce to an
array if the flag is specified multiple times. This keeps the
functionality the same regaurdless of how it is passed