feat: added dockerNetwork config option
adds a dockerNetwork policy to allow the user to define which Docker network
to use when the image is being built. You would use it alongside the standard
plugin configuration - and by default, the value of dockerNetwork is "default", as
this is the default value for the --network flag for docker build.

fixes #29