feat(plugin): add basic support for github actions
There have been several requests to gain access or over ride the build
id of the image so that the caller can control that and reference it
later in a build process. Initial attempts to expose this information
from the nextRelease property. Unfortunately, this object seems to be
sandboxed, and additional information is not carried to other plugins.
As an alternative, this adds basic support for github actions be setting
output and environment variable that include the build id and image sha
where possible

fixes: #49