fix(plugin): better support for multiple plugin instances
In the case of multiple docker plugin instances, the build ID was being
generated once causing the second image to clobber any previous builds.

This updates the build id to be generated with the compilation of the
plugin configuration and mapped to the image that its bound associated
with. The build id is generated using a sha256 of the plugin
configuration + process id and timestamp of when the process started

Fixes: #60